+++ research/nonlinearities/core/research_bert.py
@@ -2,11 +2,11 @@ from collections import OrderedDict
 
 import numpy as np
 import torch
-from lizrd.core.modules import EinMix
+from lizrd.core.misc import EinMix
 
 import lizrd.core.nn as nn
 from lizrd.core.llm import LowRank
-from lizrd.core.modules import Linear
+from lizrd.core.misc import Linear
 from lizrd.support import ash
