+++ research/reinitialization/core/linears_recycle.py
@@ -8,7 +8,7 @@ import numpy as np
 import plotly.express as px
 
 from lizrd.core.misc import get_default_device
-from lizrd.core.modules import Linear
+from lizrd.core.misc import Linear
 
 from lizrd.support.logging import (
     get_current_logger,
